langgraph_multiagent/requirements.txt
langchain-core
python-dotenv
langchain-community
ibm_watsonx_ai
langchain-ibm
langgraph
pillow